refactor: move inline route handlers to api/ modules
Extract all inline route handlers from daemon/main.ts into dedicated
api/ files, converting closure-based handlers to factory functions with
explicit dependency injection.

New files:
- api/mcp-proxy.ts: remember, supersede, plans, sessions, team handlers
- api/agent-runs.ts: run, list, get, reports, turns handlers
- api/attachments.ts: attachment file serving handler

Extended files:
- api/sessions.ts: added createSessionMutationHandlers (delete, impact)
- api/config.ts: added createPlanDirHandlers (get/update plan dirs)
- api/log-explorer.ts: added createLogIngestionHandler (external log)
- api/stats.ts: added createLiveStatsHandler (live daemon stats overlay)
